+++ src/sys/arch/namespace/impls/instr/_.rs
@@ -31,3 +31,7 @@ mod aarch64;
 #[cfg(any(target_arch = "riscv32", target_arch = "riscv64", all(doc, target_arch = "x86_64")))]
 #[cfg_attr(nightly_doc, doc(cfg(any(target_arch = "riscv32", target_arch = "riscv64"))))]
 mod riscv;
+
+#[cfg(any(target_arch = "avr", all(doc, target_arch = "x86_64")))] // WAIT:1.99:apply
+#[cfg_attr(nightly_doc, doc(cfg(target_arch = "avr")))]
+mod avr; // AVR processor instructions